Ajout de valeurs de test par Seeders

+++ database/migrations/2025_06_25_155331_create_owners_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->timestamps();
 
             // Clé étrangère
-            $table->unsignedInteger('user_id')->nullable()->unique();
+            $table->unsignedBigInteger('user_id')->nullable()->unique();
             $table->foreign('user_id')->references('user_id')->on('users');
         });
     }
